+++ Release Notes.md
@@ -7,6 +7,7 @@
 ### Bug Fixes
 
 * Fixed back button not working when launched app from today widget, 3D touch quick actions, or handoff.
+* Fixed table view highlighting for iPad.
 * Fixed today widget title being cut off in iPhone 5s or lower by reducing the name.
 * Fixed today widget some localization not showing correctly.